fix: delete messages after dispatch
When a message is dispatched to a worker, it is deleted from the internal memory store. When data is received from a worker, rather than looking up the original message to identify the worker that sent it, the new message is added to the memory store. The DataProcessor then attempts to parse the directive as a URL; if that succeeds (where "succeeds" in this case means the parsed value includes a valid URL scheme component), the message/worker is treated as a detachedContent worker and the data is return via HTTP.

Fixes RHCLOUD-12336

Signed-off-by: Link Dupont <link@sub-pop.net>