refactor: Replace all instances of 'cloud.redhat.com'
'cloud.redhat.com' as a public URL is being deprecated in favor of something else. This commit removes hard-coded references to 'cloud.redhat.com' in user-facing strings and replaces it with a compile time constant: yggdrasil.Provider. This string can be set to "cloud.redhat.com" at build time to recreate the "old" behavior, or it can be set to a new value, depending on the branding discretion.

Signed-off-by: Link Dupont <link@sub-pop.net>