feat: start workers explicitly
Workers are now started and stopped based on the existence of a
configuration file in SYSCONFDIR/yggdrasil/workers. It defines a path to
the executable that will be started and managed by yggd. Conversely, the
worker can be disabled by the removal of the configuration file.

Much of the code in exec.go was lifted into a new workers.go file.
exec.go now containers lower level routines for starting and stopping
processes. It has no awareness of the concept of a "worker". workers.go
has higher-level functions that use the startProcess and stopProcess
routines to start a worker process, create a PID file, wait for exit
status, restart and backoff restart delays, etc.

The sample echo worker config file has been updated to include two new
fields: protocol and env. protocol must be set to a supported RPC
protocol (currently "grpc" is the only supported RPC protocol). Any
strings in env must be in the form of a "VAR=VALUE" string, and are
inserted into the worker process's environment before starting. It is
forbidden to set PATH or any variable beginning with YGG_ in a worker's
configuration file. Any variables by these names will be omitted when
starting the worker process.

Signed-off-by: Link Dupont <link@sub-pop.net>