refactor!: transport interface implementation
- Rename Transport interface to Transporter and expand on the methods
  necessary to implement it.
- Refactor HTTP and MQTT transport implementations to implement
  Transporter interface.
- Move Transporter, MQTT and HTTP transports into a 'transport' package.

The Transporter interface and the concrete transport types MQTT and HTTP
do not handle yggdrasil message types; they work exclusively with byte
slices. This makes for a nice boundary between the data transport
interface and the higher-level data handling. It is even conceivable
that the Transporter interface and either the MQTT or the HTTP transport
could be used on their own to send and/or receive data, so they are put
into a separate package.

- Move the HTTP Client into an http package; it is conceivable that it
  too has value as a standalone HTTP client.
- Add a 'Client' type to yggd that encapsulates a lot of the yggdrasil
  message handling and handoff between the transport and the RPC
  dispatcher. This client is responsible for handling control messages
  appropriately, given the transport with which it has been configured.

BREAKING CHANGE: Remove the '--broker' command-line option in favor of
two options: '--protocol' to select the transport protocol (either
'mqtt' or 'http') and '--server' to identify the broker URI or the base
URI of the HTTP API server. It is no longer possible to configure a
client with multiple brokers. This breaks existing configuration files.
The 'broker' field is no longer supported.

Signed-off-by: Link Dupont <link@sub-pop.net>