NotifyEven worker service operation (#54)
* feat: add an RPC event notification method

Extend the worker RPC protocol with a "NotifyEvent" method. This method is
called by yggd when certain events occur within the program. The protocol
defines 3 events: RECEIVED_DISCONNECT, UNEXPECTED_DISCONNECT, and
CONNECTION_RESTORED. When a "disconnect" command is received by yggdrasil over
the control plane, each registered worker is sent a RECEIVED_DISCONNECT event
via the NotifyEvent method call. Workers are expected to use this notification
to gracefully shutdown and perform any cleanup necessary in the event of a
deliberate disconnection.

Notably absent from this commit is implementation supporting the
UNEXPECTED_DISCONNECT and CONNECTION_RESTORED events.

Co-authored-by: Danielle Barda <dbarda@redhat.com>
Co-authored-by: Jakub Dzon <jdzon@redhat.com>
Co-authored-by: Link Dupont <link@sub-pop.net>

Signed-off-by: Link Dupont <link@sub-pop.net>