feat: add --exclude-worker config option
Any strings listed in this argument will be stored in an exclusion list
at run time. If a worker's *basename* matches a string in this list, it
is excluded from execution when starting worker processes. For example,
`--exclude-worker echo-worker` will prevent the file
`/usr/libexec/yggdrasil/echo-worker` from being executed.

Signed-off-by: Link Dupont <link@sub-pop.net>