fix: HTTP Transport return all responses
I discovered that when the HTTP response is >400 the error is not nil,
and the transport didn't return the http.Response back to the worker.

With this commit, in case of the transport.Send fails, it returns an
error and also the http response to the worker that requested it.

Discovered on: https://github.com/project-flotta/flotta-operator/pull/144

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>