feat: add benchmark subcommand
Running `yggctl benchmark` enables a throughput test, tracking messages
published to MQTT destined for a specific worker. It calculates the time
between publishing the message and receiving the message in response. It
assumes the echo worker is available and active on the system, and uses
the "ResponseTo" field of a Data message to trace the reply.

Signed-off-by: Link Dupont <link@sub-pop.net>