feat: add events to transport
Add an event channel and handler function for the transports to emit
when network connections are lost and regained. The client sets an event
handler function that, when an event is received from the transport,
emits an event on the Dispatcher, relaying the event via a D-Bus signal
to workers.

Signed-off-by: Link Dupont <link@sub-pop.net>