fix: initialize DefaultConfig.PathPrefix
Initialize DefaultConfig.PathPrefix with a default value. It is possible
config.DefaultConfig can be used without parsing command-line arguments.
In that case, the default values that are set by the cli.Flag structs
never get applied to the DefaultConfig variable. This commit ensures
that DefaultConfig.PathPrefix always contains something (in this case,
it still defaults to "yggdrasil").

Signed-off-by: Link Dupont <link@sub-pop.net>