feat: multiple brokers
The MQTT transport supports the notion of multiple brokers; while the
HTTP transport does not, we can still support both by converting the
--server flag into a string slice. The HTTP transport is created with
only the first element of the string slice flag, while the MQTT
transport accepts the entire slice. Each broker in the slice is added
during the connection process of the MQTT transport.

Signed-off-by: Link Dupont <link@sub-pop.net>