feat: add Packit integration
Add a basic configuration for building PRs using Packit, using EL 8/9
and supported Fedora's; also, make use of our custom Copr as base for
the content.

The actions perform a similar, although easier, steps of what is done
with "make srpm" (which is what Copr does): this way we can avoid the
package installations, and we can copy all the files needed.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>