feat(yggd): Set env variables from config file
A worker may optionally install a configuration file into
/etc/yggdrasil/workers. This configuration file, if present when a
worker is started, is loaded. The 'env' field of the configuration file
is parsed as a string slice. Elements of this slice must be in the form
of a "VAR=VALUE" string, and are inserted into the worker process's
environment before starting. It is forbidden to set PATH or any variable
beginning with YGG_ in a worker's configuration file. Any variables by
these names will be omitted when starting the worker process.

HTTP proxy environment variables are automatically read from the yggd
environment and are passed into the worker environment automatically.

A sample echo worker config file has been added to demonstrate the new
env field.

Much of the code in exec.go was lifted into a new workers.go file.
exec.go now containers lower level routines for starting and stopping
processes. It has no awareness of the concept of a "worker". workers.go
has higher-level functions that use the startProcess and stopProcess
routines to start a worker process, create a PID file, wait for exit
status, restart and backoff restart delays, etc.

Signed-off-by: Link Dupont <link@sub-pop.net>