feat: Optional non-root daemon user
Two new build options, 'user' and 'group' can be used to configure a
non-root user account to run the system yggdrasil service. The value
of both of these options default to 'yggdrasil'. This value is used
to customize the D-Bus policy, systemd unit User= and Group=
directive values, and a systemd-sysuser template.

Signed-off-by: Link Dupont <link@sub-pop.net>