refactor: Don’t autostart workers
The worker discovery goroutine gets the “Features” property of each worker it discovers. This has a side effect of starting each bus-activatable worker if it was not started already. Instead of rapidly spawning each discovered worker just to probe the “Features” property, check first if the worker is running. A worker is assumed to be running if the result of ‘org.freedesktop.DBus.NameHasOwner’ is true. In this case, get the Features property and include it in the directives map; otherwise include the worker with an empty features table.